Update Store.cs
There's no evidence about a wrong deserialization of the stored informations, due to a potentially missing external libraries or a field which cannot be deserialized properly.
In case of an exception, the navigation stack  is cleared without any clue about it.
I created a Dialog, and I put an unserializable private field into it... I spent 2 hours in order to discover why the dialog were lost. I have to  download the botbuilder source code, and use it in order to undesrtand what's wrong.